+++ opentelemetry-sdk/tests/trace/export/test_export.py
@@ -175,7 +175,7 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
             OTEL_BSP_EXPORT_TIMEOUT: "4",
         },
     )
-    def test_batch_span_processor_environment_variables(self):
+    def test_args_env_var(self):
 
         batch_span_processor = export.BatchSpanProcessor(
             MySpanExporter(destination=[])
@@ -185,6 +185,37 @@ class TestBatchSpanProcessor(ConcurrencyTestBase):
         self.assertEqual(batch_span_processor.schedule_delay_millis, 2)
         self.assertEqual(batch_span_processor.max_export_batch_size, 3)
         self.assertEqual(batch_span_processor.export_timeout_millis, 4)
+
+    def test_args_env_var_defaults(self):
+
+        batch_span_processor = export.BatchSpanProcessor(
+            MySpanExporter(destination=[])
+        )
+
+        self.assertEqual(batch_span_processor.max_queue_size, 2048)
+        self.assertEqual(batch_span_processor.schedule_delay_millis, 5000)
+        self.assertEqual(batch_span_processor.max_export_batch_size, 512)
+        self.assertEqual(batch_span_processor.export_timeout_millis, 30000)
+
+    @mock.patch.dict(
+        "os.environ",
+        {
+            OTEL_BSP_MAX_QUEUE_SIZE: "a",
+            OTEL_BSP_SCHEDULE_DELAY: " ",
+            OTEL_BSP_MAX_EXPORT_BATCH_SIZE: "One",
+            OTEL_BSP_EXPORT_TIMEOUT: "@",
+        },
+    )
+    def test_args_env_var_value_error(self):
+
+        batch_span_processor = export.BatchSpanProcessor(
+            MySpanExporter(destination=[])
+        )
+
+        self.assertEqual(batch_span_processor.max_queue_size, 2048)
+        self.assertEqual(batch_span_processor.schedule_delay_millis, 5000)
+        self.assertEqual(batch_span_processor.max_export_batch_size, 512)
+        self.assertEqual(batch_span_processor.export_timeout_millis, 30000)
 
     def test_on_start_accepts_parent_context(self):
         # pylint: disable=no-self-use